+++ backend/src/main/java/com/entelgy/project/demo/Service/UserService.java
@@ -32,8 +32,12 @@ public class UserService {
         return null;
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     public List<User> getAllUsers() {
